Moved heatmap to it's own component and added it to the Patient Dashboard Page
Heatmap had been on the example page Weather while initially setting it up and testing it.
Have now moved it to a separate UI Component so it can be added to different pages easily.
This also involved changing the way the heatmap was being refreshed by making it a background task.
Added the Heatmap component to the Patient Dashboard page.